debt: move or remove `NullXYZServices`
We have a few NullXYZServices. Those are useful for testing but we should move them to separate files so that we don't end up shipping them with the product
