+++ src/vs/editor/test/browser/services/openerService.test.ts
@@ -7,7 +7,8 @@ import { Disposable } from 'vs/base/common/lifecycle';
 import { URI } from 'vs/base/common/uri';
 import { OpenerService } from 'vs/editor/browser/services/openerService';
 import { TestCodeEditorService } from 'vs/editor/test/browser/editorTestServices';
-import { CommandsRegistry, ICommandService, NullCommandService } from 'vs/platform/commands/common/commands';
+import { CommandsRegistry, ICommandService } from 'vs/platform/commands/common/commands';
+import { NullCommandService } from 'vs/platform/commands/test/common/nullCommandService';
 import { ITextEditorOptions } from 'vs/platform/editor/common/editor';
 import { matchesScheme, matchesSomeScheme } from 'vs/platform/opener/common/opener';
 import { TestThemeService } from 'vs/platform/theme/test/common/testThemeService';
